logistream/requirements.txt
pyspark==3.5.1
delta-spark==3.1.0
kafka-python-ng==2.2.3
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
apache-airflow==2.9.1
pandas==2.2.2
pyarrow==16.0.0
pytest==8.2.0
httpx==0.27.0
faker==25.1.0
